remove a weird special case in path traversal
Previously, given

  my $c = container 'Foo', as {
      container 'Bar', as {
          service 'Baz' => (value => 1);
      };
  };

you could do $c->fetch('/Foo/Foo/Bar/Bar/Bar/Baz') and still get out the
Baz service, because giving the name of the current container as a path
component was treated as a no-op. This doesn't actually make any sense,
is fairly confusing, and required extra code to implement it as a
special case, so I'm not sure why it was behaving that way at all
originally. So, remove it.